+++ changelog.md
@@ -11,7 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Added
 
 - Doc only: Systematically mention all crate saftey risks (E.G. caller supplied URLs) (`f1ce800`)
-- Define complexity limits on attacker-controlled scene generation (``)
+- Define complexity limits on attacker-controlled scene generation (`38b3518`)
+- Harden CI against compromised GitHub Actions (``)
 
 ## Fixed